+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - (cli) [#12](https://github.com/evmos/proposer/pull/12) Use [clap](https://github.com/clap-rs/clap) for CLI handling.
 - (all) [#13](https://github.com/evmos/proposer/pull/13) Implement better error handling using [thiserror](https://github.com/dtolnay/thiserror).
 - (tests) [#11](https://github.com/evmos/proposer/pull/11) Mock API responses for tests.
+- (block) [#15](https://github.com/evmos/proposer/pull/15) Add error handling for blocks data.
 
 ## Legacy Changelog from [Original Repo](https://github.com/MalteHerrmann/upgrade-helper)